Removing ForgotPasswordSpi, moving Provider and Factory to a better location
