+++ services/src/main/resources/META-INF/services/org.keycloak.services.resources.spi.RealmResourceProviderFactory
@@ -1,1 +1,1 @@
-org.keycloak.protocol.ForgotPasswordProviderFactory
+org.keycloak.services.forgotpassword.ForgotPasswordProviderFactory
